# docs/normative/v0.6/contracts/AI-05_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Paquetería Local Culiacán API
  version: 0.6.0
  description: Contrato REST normativo v0.6. REST/PostgreSQL son autoridad; SignalR distribuye eventos confirmados. Seguridad
    tenant, tracking público y outbox siguen contratos ejecutables.
servers:
- url: /api/v1
security:
- oidc: []
...
paths:
  /quotes:
    post:
      tags:
      - Quotes
      operationId: createQuote
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
      responses:
        '201':
          description: Quote created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '422':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /quotes/{quoteId}:
    get:
      tags:
      - Pricing
      operationId: getQuote
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - name: quoteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    OrganizationContext:
      name: X-Organization-Id
      in: header
      required: true
      description: Organización activa; debe corresponder a una membresía ACTIVE del usuario.
      schema:
        type: string
        format: uuid
  responses:
...
    ValidationError:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated actor lacks capability inside an otherwise visible tenant context
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UniformNotFound:
      description: Missing, cross-tenant, expired or revoked resource. Response must not reveal existence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ServiceUnavailable:
      description: >-
        Required provider or infrastructure dependency is unavailable. Every
        authenticated operation can report it while the identity context
        cannot be resolved, and every tenant operation while the mandatory
        PLATFORM_ADMIN tenant-activation audit cannot be written
        (AI05-DECLARE-EMITTED-ERRORS).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Money:
      type: object
      required:
      - currency
      - amount_cents
      properties:
        currency:
          type: string
          const: MXN
        amount_cents:
          type: integer
          minimum: 0
          format: int64
...
    AddressInput:
      type: object
      required:
      - address_text
      - contact_name
      - phone
      properties:
        address_text:
          type: string
          minLength: 8
        contact_name:
          type: string
        phone:
          type: string
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        references:
          type: string
          maxLength: 500
    CreateQuoteRequest:
      type: object
      required:
      - origin
      - destination
      - service_type
      - packages
      properties:
        client_account_id:
          type: string
          format: uuid
        origin:
          $ref: '#/components/schemas/AddressInput'
        destination:
          $ref: '#/components/schemas/AddressInput'
        service_type:
          type: string
          enum:
          - SAME_DAY
          - URGENT
          - SCHEDULED_ROUTE
        consolidated_route:
          type: boolean
          default: false
        packages:
          type: array
          minItems: 1
          maxItems: 20
          description: AI05-INPUT-LIMITS; more than 20 packages is a validation error (422).
          items:
            $ref: '#/components/schemas/PackageInput'
      description: Owner organization is derived from X-Organization-Id after membership validation.
    PackageInput:
      type: object
      required:
      - description
      - weight_grams
      - declared_value_cents
      properties:
        description:
          type: string
          maxLength: 250
        weight_grams:
          type: integer
          minimum: 1
        declared_value_cents:
          type: integer
          minimum: 0
          format: int64
        length_mm:
          type: integer
          minimum: 1
        width_mm:
          type: integer
          minimum: 1
        height_mm:
          type: integer
          minimum: 1
    Quote:
      type: object
      required:
      - id
      - net
      - tax
      - total
      - rule_ids
      - expires_at
      - origin_location_id
      - destination_location_id
      - service_type
      - consolidated_route
      - package_snapshot
      - pricing_tier
      - minimum_total_cents_snapshot
      - pricing_policy_version
      - status
      - city_id
      properties:
        id:
          type: string
          format: uuid
        net:
          $ref: '#/components/schemas/Money'
        tax:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        rule_ids:
          type: array
          items:
            type: string
            format: uuid
        breakdown:
          type: array
          items:
            type: object
            additionalProperties: true
        expires_at:
          type: string
          format: date-time
        origin_location_id:
          type: string
          format: uuid
        destination_location_id:
          type: string
          format: uuid
        service_type:
          type: string
          enum:
          - SAME_DAY
          - URGENT
          - SCHEDULED_ROUTE
        consolidated_route:
          type: boolean
        package_snapshot:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PackageInput'
        city_id:
          type: string
          format: uuid
        service_area_id:
          type:
          - string
          - 'null'
          format: uuid
        pricing_tier:
          $ref: '#/components/schemas/PricingTier'
        minimum_total_cents_snapshot:
          type: integer
          minimum: 0
          format: int64
        pricing_policy_version:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - USED
          - EXPIRED
          - REVOKED
        request_snapshot_redacted:
          type: object
          additionalProperties: true
          description: Must not contain plaintext address, contact name, telephone or access notes.
...
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...
    PricingTier:
      type: string
      enum:
      - OCCASIONAL
      - BUSINESS_1_49
      - BUSINESS_50_199
      - BUSINESS_200_499
      - BUSINESS_500_PLUS
      - CUSTOM
...